requirements.txt
SQLAlchemy==1.4.23
greenlet==1.1.1
django-environ==0.6.0
PyMySQL==1.0.2